fix(github): Move endpoint to press.api.bench
(cherry picked from commit 01fd4013591fc224119657b1703d6b1764aadaae)
